edit

+++ JpaTicketManager/src/main/java/com/mykheikin/jpaproject/service/RoleService.java
@@ -10,5 +10,5 @@ public interface RoleService {
 
 	Role findByName(String name);
 
-	Role findById(int id);
+	Role findById(Long id);
 }
